Move concatenate preprocessor function to its own module
